+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 Features
   * :issue:`233`: Add config endpoint for Know Me app. It contains information such as the lowest useable iOS app version.
   * :issue:`259`: Add optional profile image for users.
+  * :issue:`263,277`: Automatically create a Know Me user for each user. The Know Me user's image defaults to the user's profile image.
   * :issue:`299`: Add additional information to media resources. The resources can have a link instead of a file, and they have an integer to hint how they should be styled.